requirements.txt
gradio
spacy
PyPDF2
pillow
pytesseract
sentence-transformers
faiss-cpu
numpy
mysql-connector-python
python-dotenv